Display unread notification count on notification icon
