--- app/controllers/test/components/notifications_icons_controller.rb
@@ -1,9 +1,0 @@
-class Test::Components::NotificationsIconsController < Test::BaseController
-  def show
-    @user = User.first
-  end
-
-  def update
-    NotificationsChannel.broadcast_changed(User.first, count: params[:count].to_i)
-  end
-end
