--- app/views/test/components/notifications_icons/show.html.haml
@@ -1,9 +1,0 @@
-%h1 Notifications
-%h2 Icon
-= form_with method: :patch do |f|
-  = f.label :count
-  = f.number_field :count
-  = f.submit "Update"
-
-#test-area
-  = ReactComponents::Common::NotificationsIcon.new(@user)
